added atof

+++ ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include <stddef.h>
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	int	i;
